// frontend/src/api/generated/models/importFileType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type ImportFileType = typeof ImportFileType[keyof typeof ImportFileType];


export const ImportFileType = {
  GoogleSheetsExportCsv: 'GoogleSheetsExportCsv',
  StandardCsvTemplate: 'StandardCsvTemplate',
  GarminActivityExport: 'GarminActivityExport',
  MySasyExport: 'MySasyExport',
} as const;