// frontend/src/api/generated/models/importRowStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type ImportRowStatus = typeof ImportRowStatus[keyof typeof ImportRowStatus];


export const ImportRowStatus = {
  Valid: 'Valid',
  Warning: 'Warning',
  Error: 'Error',
  DuplicateSkipped: 'DuplicateSkipped',
} as const;